Break up utils into methods/breakpoints.
