+++ src/rules/border-colors.ts
@@ -1,5 +1,5 @@
 import { RuleFn } from "../config";
-import { newMethodsForProp } from "../utils";
+import { newMethodsForProp } from "../methods";
 
 export const borderColorRules: RuleFn = ({ palette }) => {
   const defs = Object.fromEntries(
