+++ src/rules/border-radius.ts
@@ -1,5 +1,5 @@
 import { RuleFn } from "../config";
-import { newMethodsForProp } from "../utils";
+import { newMethodsForProp } from "../methods";
 
 export const borderRadiusRules: RuleFn = () =>
   newMethodsForProp("borderRadius", {
