+++ src/rules/coordinates.ts
@@ -1,4 +1,4 @@
-import { newIncrementMethods, Prop } from "../utils";
+import { newIncrementMethods, Prop } from "../methods";
 import { RuleFn } from "../config";
 
 const directions: Prop[] = ["top", "right", "bottom", "left"];
