+++ src/rules/floats.ts
@@ -1,5 +1,5 @@
 import { RuleFn } from "../config";
-import { newMethodsForProp } from "../utils";
+import { newMethodsForProp } from "../methods";
 
 // https://github.com/tachyons-css/tachyons/blob/master/src/_floats.css
 export const floatRules: RuleFn = () =>
