+++ src/rules/text-decoration.ts
@@ -1,4 +1,4 @@
-import { newMethodsForProp } from "../utils";
+import { newMethodsForProp } from "../methods";
 import { RuleFn } from "../config";
 
 export const textDecorationRules: RuleFn = () =>
