+++ src/rules/zIndex.ts
@@ -1,4 +1,4 @@
-import { newMethodsForProp, newParamMethod } from "../utils";
+import { newMethodsForProp, newParamMethod } from "../methods";
 import { RuleFn } from "../config";
 
 // https://github.com/tachyons-css/tachyons-z-index/blob/master/src/tachyons-z-index.css
